Reprise à 0 du projet : - Vérification d'un mot dans la grille revue - Commentaire pour génération XML commencés

+++ mots.h
@@ -8,10 +8,10 @@
 #include <string.h>
 #include <ctype.h>
 
-int Nb_de_lettres_grille(char grille[8][8], char lettre_a_compter, short longueur);
+int Nb_de_lettres_grille(char grille[8][8], char lettre_a_compter, int longueur);
 void Coordonnees_lettre(char lettre, int longueur, char grille[8][8], int *indiceL, int *indiceC);
 void Obtention_lettres_autour(char lettre_autour[8], int indiceC, int indiceL, char grille[8][8], int longueur);
 int Position_lettre_tab_lettre_autour(char lettre_autour[8], char lettre, int longueur);
-int Traitement_mot(char tabmots[], char grille[8][8], short longueur);
+int Traitement_mot(char tabmots[], char grille[8][8], int longueur);
 int Verification_francais(char mot_a_comparer[]);
-void Saisie_de_mots(short temps_limite, char grille[8][8], short longueur, char tabmots[][26]);
+void Saisie_de_mots(int temps_limite, char grille[8][8], int longueur, char tabmots[][26]);
